Move qgsauthorizationsettings.h to core/auth
