+++ src/core/auth/qgsauthorizationsettings.h
